pci/util: add utility function to get device vfio id
The function to get the device vfio id does not depend on any
iommufd/vfio functionality. Move it to pci/util.
